Query levantamentos TARV

calc_arv_pickups() devolve uma listagem dos levantamentos ARV feito during um período especificado pelo utilizador

Usage

calc_arv_pickups(
  con,
  enddate,
  opendate,
  arv_pickup_source = "FILA",
  filter_by_location = FALSE,
  location_uuid = "4be5f1a9-832c-4717-be41-ef4b6311c0ef"
)

Arguments

con

Ligação à base de dados MozART 2.0

enddate

Data do fecho do período (introduzir como “AAAA-MM-DD”)

opendate

Data da abertura do período (introduzir como “AAAA-MM-DD”)

arv_pickup_source

Fonte de levantamento de ARV. Opções são "FILA", "Reception", "Any". Default é "FILA

"FILA" por defeito, devolve levantamentos documentados na FILA ou fonte alternativa da farmácia

"Reception" devolve levantamentos documentados na Ficha de Seguimento, Mapa dos Levantamentos

"Any" devolve todos levantamentos

filter_by_location

FALSE por defeito

TRUE executa sobre uma US específica

FALSE executa sobre todas as US contidas no MozART 2.0

location_uuid

location_uuid da US a filtrar quando filter_by_location é definido como TRUE

Value

Um quadro de dados contendo uma listagem individual dos levantamentos TARV

Examples