Detectar transição etária

project_age_out() devolve uma listagem dos clientes clientes com idades projectadas para uma data definida pelo utilizador, bem como colunas para indicar se fizeram a transição da idade pediátrica para a idade adulta e se a sua faixa etária de 5 anos mudou

Usage

project_age_out(df, ref_date)

Arguments

df

Quadro de dados contendo a data de nasicmento do cliente

ref_date

Data de referência para calcular a idade projectada do cliente

Value

Quadro de dados com variáveis idade, idade projectada, faixa etária, faixa etária projectada, e colunas que indicam casos de transição etária

Examples