Query uuids dos activos no TARV

pull_active_uuid() devolve um vector dos uuid's dos clientes activos em TARV a partir de uma data definida pelo utilizador

Usage

pull_active_uuid(
  con,
  enddate,
  defaulter_tolerance = 28,
  filter_by_location = FALSE,
  location_uuid = "4be5f1a9-832c-4717-be41-ef4b6311c0ef"
)

Arguments

con

Ligação à base de dados MozART 2.0

enddate

Data de fecho do período (introduzir como “AAAA-MM-DD”)

defaulter_tolerance

Número de dias de tolerância antes dos faltosos serem considerados inactivos (abandonos)

filter_by_location

Lógico. Se TRUE, o query corre sobre uma unidade sanitária específica. Se FALSE, o query corre sobre todas as unidade sanitária contidas na base de dados

location_uuid

location_uuid da unidade sanitária a filtrar quando filter_by_location é definido como TRUE

Value

pull_active_uuid devolve um vector contendo os patient_uuid unicos dos pacientes activos em TARV

Examples

if (FALSE) { # \dontrun{

#' Caso de uso simples executado sobre uma US, tolerância de 28 dias
 df <- pull_active_uuid(
          con,
          enddate = '2024-09-20',
          defaulter_tolerance = 28,
          filter_by_location = TRUE,
          location_uuid = location_meripo)} # }