Recodificar idade

recode_age() devolve a idade e a faixa etária do cliente

Usage

recode_age(df, age_column = "age", variable_name = "age_band")

Arguments

df

Quadro de dados contendo a variável a recodificar

age_column

Variável no quadro de dados introduzido contendo a idade do cliente

variable_name

Nome da variável que será criada no quadro de dados com a faixa etária do client (default = 'age_band')

Value

recode_age devolve um quadro de dados com a coluna idade recodificada

Examples

if (FALSE) { # \dontrun{

 df <- recode_age(
          df = df,
          age_column = "age")} # }